# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "mpc-ext server",
    "description": "Holds one share of a 2-of-3 MPC key, joins signing, and verifies server-origin WebAuthn passkey ceremonies.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/v1/passkeys/handoff": {
      "post": {
        "tags": [
          "passkey"
        ],
        "summary": "Creates the one-use browser handoff after verifying device possession.",
        "operationId": "handoff",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HandoffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "browser ceremony handoff",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HandoffResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HandoffRequest": {
        "type": "object",
        "description": "Starts a server-origin browser ceremony after device-key authentication.",
        "required": [
          "wallet_id",
          "purpose"
        ],
        "properties": {
          "digest": {
            "type": [
              "string",
              "null"
            ]
          },
          "operation_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "purpose": {
            "type": "string",
            "description": "`register`, `sign`, or `recovery`."
          },
          "wallet_id": {
            "type": "string"
          }
        }
      },
      "HandoffResponse": {
        "type": "object",
        "description": "A one-use token delivered to the extension and then submitted in an HTML form body.",
        "required": [
          "ceremony_id",
          "handoff_token"
        ],
        "properties": {
          "ceremony_id": {
            "type": "string"
          },
          "handoff_token": {
            "type": "string"
          }
        }
      },
...
    }
  }
}